test: collapse ATS tests into a single tagged file
The cluster-reachability check now also runs in the upgrade scenario;
it was smoke-only as a leftover from the pytest markers.
